requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
beartype==0.19.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.115.6
fire==0.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.57.4
packaging==24.2
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
PyDrive==1.3.1
pyparsing==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
termcolor==2.4.0
tqdm==4.67.1
typing_extensions==4.12.2
unityparser==4.0.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.1.3
